pyproject.toml
[tool.poetry]
name = "langgraph-engineer"
version = "0.0.1"
description = "Hello, LangGraph Engineer!"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langgraph = "^0.2.5"
langchain-anthropic = "^0.1.23"
langchain-core = "^0.2.33"
langchain-openai = "^0.1.22"
requests = "^2.32.3"
openai = "^1.42.0"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
